+++ src/confusius/bids/__init__.py
@@ -32,9 +32,9 @@ from confusius.bids.coordinates import (
     create_bids_slice_timing_from_coordinate,
     create_slice_time_coordinate_from_bids,
 )
-from confusius.bids.events import DEFAULT_TRIAL_TYPE, read_events, write_events
+from confusius.bids.events import DEFAULT_TRIAL_TYPE, load_events, save_events
 from confusius.bids.mapping import EXPLICIT_BIDS_FIELD_MAPPINGS, from_bids, to_bids
-from confusius.bids.physio import read_physio
+from confusius.bids.physio import load_physio
 from confusius.bids.validation import (
     FUSI_BIDS_FIELDS,
     FUSIBIDSMetadata,
@@ -57,8 +57,8 @@ __all__ = [
     "create_slice_time_coordinate_from_bids",
     # Events
     "DEFAULT_TRIAL_TYPE",
-    "read_events",
-    "write_events",
+    "load_events",
+    "save_events",
     # Physio
-    "read_physio",
+    "load_physio",
 ]